requirements/dev.txt
-r base.txt
django-rosetta==0.10.3
polib==1.2.0
# para desenvolvimento e debugging